Release: io7m-smfj 0.4.0
Code new: Add schema-set and schema-check commands. (tickets: https://github.com/io7m/smf/issues/19)
Code new: Add trim command. (tickets: https://github.com/io7m/smf/issues/21)
Code change: Update the random access parser API to deliver metadata. (tickets: https://github.com/io7m/smf/issues/23)
Code new: Add an adapter that can deliver parser events from an SMFMemoryMesh. (tickets: https://github.com/io7m/smf/issues/18)
Code change: Add an optional exception to parse and processing errors. (tickets: https://github.com/io7m/smf/issues/22)
Code fix: Fix a major issue with 8-bit data not being serialized. (tickets: https://github.com/io7m/smf/issues/24)
Code change: Make the SMFMetadata vendor/schema fields into long values to ease the use of them as unsigned 32 bit values.

+++ io7m-smfj-format-text/src/main/java/com/io7m/smfj/format/text/SMFTV1Serializer.java
@@ -700,8 +700,8 @@ final class SMFTV1Serializer implements SMFSerializerType
 
   @Override
   public void serializeMetadata(
-    final int vendor,
-    final int schema,
+    final long vendor,
+    final long schema,
     final byte[] data)
     throws IOException, IllegalStateException
   {
@@ -716,8 +716,8 @@ final class SMFTV1Serializer implements SMFSerializerType
       this.writer.append(
         String.format(
           "meta %s %s %d",
-          Integer.toUnsignedString(vendor, 16),
-          Integer.toUnsignedString(schema, 16),
+          Long.toUnsignedString(vendor, 16),
+          Long.toUnsignedString(schema, 16),
           Integer.valueOf(lines.size())));
       this.writer.newLine();
